feat(results): Create base datasource for results (#136)
