+++ src/datasources/results/ResultsDataSourceBase.ts
@@ -1,0 +1,15 @@
+import { DataSourceBase } from "core/DataSourceBase";
+import { ResultsQuery } from "./types";
+import { DataQueryRequest, DataFrameDTO, TestDataSourceResponse } from "@grafana/data";
+
+export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery> {
+  baseUrl = this.instanceSettings.url + '/nitestmonitor';
+
+  abstract runQuery(query: ResultsQuery, options: DataQueryRequest): Promise<DataFrameDTO>;
+
+  abstract shouldRunQuery(query: ResultsQuery): boolean;
+
+  testDatasource(): Promise<TestDataSourceResponse> {
+    throw new Error("Method not implemented.");
+  }
+}
